Try not to depend on import imagesize, use known locations first, with checking
Conflicts:
	cajparser.py

+++ cajparser.py
@@ -1,6 +1,5 @@
 import os
 import struct
-import imagesize
 from shutil import copy
 from subprocess import check_output, STDOUT, CalledProcessError
 from utils import fnd, fnd_all, add_outlines, fnd_rvrs, fnd_unuse_no
@@ -383,12 +382,18 @@ class CAJParser(object):
                         0
                     )
                 elif (image_type[image_type_enum] == "JPEG"):
-                    #(height, width) = struct.unpack(">HH", image_data[163:167])
-                    with open(".tmp.jpg", "wb") as f:
-                        f.write(image_data)
-                    (width, height) = imagesize.get(".tmp.jpg")
-                    os.remove(".tmp.jpg")
-
+                    # stock libjpeg location
+                    (SOFn, frame_length, bits_per_pixel, height, width) = struct.unpack(">HHBHH", image_data[158:167])
+                    if (SOFn != 0xFFC0):
+                        # "Intel(R) JPEG Library" location
+                        (SOFn, frame_length, bits_per_pixel, height, width) = struct.unpack(">HHBHH", image_data[0x272:0x27b])
+                        if (SOFn != 0xFFC0):
+                            # neither works, try brute-force
+                            import imagesize
+                            with open(".tmp.jpg", "wb") as f:
+                                f.write(image_data)
+                                (width, height) = imagesize.get(".tmp.jpg")
+                                os.remove(".tmp.jpg")
                     if (image_type_enum == 1):
                         # non-inverted JPEG Images
                         height = -height
